requirements.txt
deepface==0.0.89
opencv_python==4.9.0.80
pyserial==3.5
